+++ l1vm-build-wspace.sh
@@ -47,5 +47,10 @@ fi
 #cleanup
 rm out.l1com
 rm out.l1obj
+
+if test -e "out.md"; then
 rm out.md
+fi
+if test -e "tmp.md"; then
 rm tmp.md
+fi
